requirements.txt
pytz
pyephem
Adafruit-BMP
Adafruit-DHT
Adafruit-GPIO
Adafruit-PureIO
influxdb
picamera
requests
